GUnpackDXT: fix pointer truncations

+++ Versions/Temporary/Engine/Sources/Image/GUnpackDXT.cpp
@@ -391,7 +391,7 @@ void DecompressDXT1( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 	for( j = 0; j < yblocks; ++j )
 	{
 		// 8 bytes per block
-		pBlock = (SDXTColBlock*) ( (DWORD)pCompBytes + j * xblocks * 8 );
+		pBlock = (SDXTColBlock*) ( (uintptr_t)pCompBytes + j * xblocks * 8 );
 
 
 		for( i=0; i < xblocks; i++, pBlock++ )
@@ -404,7 +404,7 @@ void DecompressDXT1( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 			// now decode the color block into the bitmap bits
 			// inline func:
 
-			pImPos = (DWORD*)((DWORD)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
+			pImPos = (DWORD*)((uintptr_t)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
 
 
 			DecodeColorBlock( pImPos, pBlock, hdr.dwWidth, (DWORD*)&col_0, (DWORD*)&col_1,
@@ -445,7 +445,7 @@ void DecompressDXT3( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 		// 8 bytes per block
 		// 1 block for alpha, 1 block for color
 
-		pBlock = (SDXTColBlock*) ( (DWORD)pCompBytes + j * xblocks * 16 );
+		pBlock = (SDXTColBlock*) ( (uintptr_t)pCompBytes + j * xblocks * 16 );
 
 		for ( i = 0; i < xblocks; i++, pBlock++ )
 		{
@@ -463,7 +463,7 @@ void DecompressDXT3( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 			// Decode the color block into the bitmap bits
 			// inline func:
 
-			pImPos = (DWORD*)((DWORD)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
+			pImPos = (DWORD*)((uintptr_t)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
 
 
 			DecodeColorBlock( pImPos, pBlock, hdr.dwWidth, (DWORD*)&col_0, (DWORD*)&col_1,
@@ -515,7 +515,7 @@ void DecompressDXT5( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 		// 8 bytes per block
 		// 1 block for alpha, 1 block for color
 
-		pBlock = (SDXTColBlock*) ( (DWORD)pCompBytes + j * xblocks * 16 );
+		pBlock = (SDXTColBlock*) ( (uintptr_t)pCompBytes + j * xblocks * 16 );
 
 		for ( i = 0; i < xblocks; i++, pBlock++ )
 		{
@@ -536,7 +536,7 @@ void DecompressDXT5( DWORD *pRes, const SDDSHeader &hdr, const BYTE *pCompBytes 
 			// Decode the color block into the bitmap bits
 			// inline func:
 
-			pImPos = (DWORD*)((DWORD)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
+			pImPos = (DWORD*)((uintptr_t)pBase + i*16 + (j*4) * hdr.dwWidth * 4 );
 
 
 			DecodeColorBlock( pImPos, pBlock, hdr.dwWidth, (DWORD*)&col_0, (DWORD*)&col_1,
